moves add_sum_of_rows to until file and adds test
